BCS

PairingHamiltonian.BCS — Method

BCS(Norb, Nocc, epsilon, gval, to, debug_mode; max_iter=1000)

Main function to calculate the ground state energy of the BCS Hamiltonian.

Arguments

Norb::Int: number of orbitals

Nocc::Int: number of occupied states

epsilon::Vector{Float64}: single-particle energies

gval::Float64: interaction strength

to::TimerOutput: timer object to measure the elapsed time

debug_mode::Int: specifying the debug mode

Optional arguments

max_iter::Int(1000): maximum number of iterations

PairingHamiltonian.E_BCS_from_uv — Method

E_BCS_from_uv(epsilon, epsilon_p, vs, Delta, lambda, gval, debug_mode)

Calculate the ground state energy of the BCS Hamiltonian from the Bogoliubov coefficients vs and us:

\[E_\mathrm{BCS} = \sum_{a} \left( \epsilon'_a v^2_a - \frac{g v^4_a}{2} \right)\]

PairingHamiltonian.Neval — Method

Neval(epsilon_p, lambda, Delta)::Float64

Calculate the expectation value of the particle number from the chemical potential lambda and the gap parameter Delta via the following equation:

\[N = \sum_{\alpha} \frac{1}{2} \left( 1 - \frac{\epsilon'_a - \lambda}{\sqrt{(\epsilon'_a - \lambda)^2 + \Delta^2}} \right)\]

PairingHamiltonian.eval_Delta — Method

eval_Delta(vs, us, gval)

Calculate the gap parameter Delta: $\Delta \equiv \frac{g}{2} \sum_{\alpha} u_a v_a$

Arguments

vs::Vector{Float64}: v[i]

us::Vector{Float64}: u[i]

gval::Float64: g

PairingHamiltonian.eval_eps_p! — Method

eval_eps_p!(epsilon_p, epsilon, vs, gval)

Calculate the epsilon_p from epsilon and vs.

Arguments

epsilon_p::Vector{Float64}: epsilon_p[i] = epsilon[i] - g v[i]^2

epsilon::Vector{Float64}: epsilon[i]

vs::Vector{Float64}: v[i]

PairingHamiltonian.reeval_Delta — Function

reeval_Delta(lambda, Delta, epsilon_p, gval, debug_mode, alpha=0.5)::Float64

Update the gap parameter Delta iteratively with the following equation:

\[\Delta^\mathrm{new} := \alpha \Delta^\mathrm{new} + (1-\alpha) \frac{g}{4} \sum_{\alpha} \frac{\Delta}{\sqrt{(\epsilon'_a - \lambda)^2 + \Delta^2}}\]

Arguments

lambda::Float64: chemical potential

Delta::Float64: current gap parameter

epsilon_p::Vector{Float64}: epsilon_p[i] = epsilon[i] - g v[i]^2

gval::Float64: g

debug_mode::Int: specify the debug mode

alpha::Float64: mixing parameter to update Delta, which may be useful to stabilize the calculation

PairingHamiltonian.update_lambda — Function

update_lambda(Nocc, lambda, Delta, epsilon_p, gval, debug_mode, alpha=0.5)

Update the chemical potential lambda iteratively with the following equation:

\[\lambda^\mathrm{new} := \frac{g}{2} \left( N_\mathrm{occ.} - \frac{N_\mathrm{orb.}}{2} + \frac{1}{2} \sum_{\alpha} \frac{\epsilon'_a }{\sqrt{(\epsilon'_a - \lambda)^2 + \Delta^2}} \right)\]

Arguments

Nocc::Int: number of occupied states

lambda::Float64: chemical potential

Delta::Float64: gap parameter

epsilon_p::Vector{Float64}: epsilon_p[i] = epsilon[i] - g v[i]^2

gval::Float64: g

debug_mode::Int: specify the debug mode

alpha::Float64: mixing parameter to update lambda

PairingHamiltonian.uv_from_Lam_Delta! — Method

uv_from_Lam_Delta!(Delta, lambda, epsilon_p::Vector{Float64}, us::Vector{Float64}, vs::Vector{Float64})

Destructively update the Bogoliubov coefficients us and vs from the gap parameter Delta and the chemical potential lambda:

\[u^2_a = \frac{1}{2} \left( 1 + \frac{\epsilon'_a - \lambda}{\sqrt{(\epsilon'_a - \lambda)^2 + \Delta^2}} \right) \\ v^2_a = \frac{1}{2} \left( 1 - \frac{\epsilon'_a - \lambda}{\sqrt{(\epsilon'_a - \lambda)^2 + \Delta^2}} \right) \]

Arguments

Delta::Float64: gap parameter

lambda::Float64: chemical potential

epsilon_p::Vector{Float64}: epsilon_p[i] = epsilon[i] - g v[i]^2

us::Vector{Float64}: u coefficients of Bogoliubov transformation

vs::Vector{Float64}: v coefficients of Bogoliubov transformation