Coupled Cluster

PairingHamiltonians.CCD — Method

CCD(F, gval, Nocc, to, debug_mode; itnum=100, tol=1e-9, mix_ratio=0.25)

Main function to calculate the ground state energy of the pairing Hamiltonian using the coupled cluster method. Since the Hamiltonian is the pairing model, we don't need to consider 1p1h excitations, CCD instead of CCSD, etc.

Arguments

F::Array{Float64, 2}: Fock matrix from the Hartree-Fock calculation

gval::Float64: strength of the pairing interaction

Nocc::Int: number of occupied states

to::TimerOutput: timer object to measure the elapsed time

debug_mode::Int: specifying the debug mode

Optional arguments

itnum::Int(100): maximum number of iterations

tol::Float64(1e-9): convergence criterion

mix_ratio::Float64(0.25): mixing ratio for the update of the t2 amplitudes

References

Shavitt, I., & Bartlett, R. J. (2009). Many-body methods in chemistry and physics: MBPT and coupled-cluster theory. Cambridge University Press.

Justin G. Lietz, Samuel Novario, Gustav R. Jansen, Gaute Hagen & Morten Hjorth-Jensen, Computational Nuclear Physics and Post Hartree-Fock Methods, part of the book series: Lecture Notes in Physics ((LNP,volume 936))

PairingHamiltonians.construct_chi — Function

i,j,k are hole indices, a,b,c are particle indices.

PairingHamiltonians.eval_χ_hhhh! — Method

\[\langle kl | \chi | ij \rangle = \langle kl | v | ij \rangle + \frac{1}{2} \sum_{cd} \langle kl | v | cd \rangle \langle cd | t | ij \rangle\]

PairingHamiltonians.get_vecidx — Method

To get continuous idx for pp or hh from i and j.

PairingHamiltonians.hash_key4 — Method

hash_key4(i, j, a, b)::UInt64

Function to encode i,j,a,b into a single UInt64. From the lowest 8 bits, b, a, j, i are stored. If you want to work with larger model space, requiring e.g. single particle states > 255, you need to modify this function.

PairingHamiltonians.unhash_key4 — Method

unhash_key4(key::UInt64)::Tuple{Int, Int, Int, Int}

Function to decode i,j,a,b from a single UInt64.

PairingHamiltonians.vecidx_to_idx — Method

To get i and j for pp or hh from a vectorized index. Note that i must be less than j.