Eigenvector Continuation

PairingHamiltonian.EC_from_FCI — Method

EC_from_FCI(Norb_in, Nocc_in; target_dirs="eigenstates_fullCI", gvals_specified::Vector{Float64} = Vector{Float64}[], gvals_target=collect(-2.0:0.1:2.0), degenerate = true, delta_eps::Float64=1.0)

Main function to compute the energy curve by EC method from the FCI wavefunctions. Assuming that the FCI wavefunctions are already evaluated and saved in a directory.

Arguments

Norb_in::Int64: Number of orbitals.

Nocc_in::Int64: Number of electrons.

target_dirs::String: Directory name where the FCI wavefunctions are saved.

gvals_specified::Vector{Float64}: List of gvals to be used as snapshots. If empty, all the snapshots extracted via glob will be used.

PairingHamiltonian.read_snapshot — Method

read_snapshot(fn::String)

Read the snapshot from the file fn and return the data as a tuple of Norb, Nocc, gval, and evec. Note that the snapshot is assumed to be written out by write_Exact_wf_hdf5 function in src/fullCI.jl.

PairingHamiltonian.solve_gen_eig — Method

solve_gen_eig(tildeH, tildeN)

Solve the generalized eigenvalue problem $\tilde{H} v = E N v$ and return the minimum eigenvalue $E$. Now it is implemented by the eigen function in Julia, but it can be ill-conditioned for large systems, so it is better to use e.g. Cholesky decomposition and iterative solver like Arpack or KrylovKit for more general cases.