Hamiltonian

Functions for constructing paring Hamiltonian and some utilities for handling it.

\[H_\mathrm{pair} = \sum_{i=1}^{N} \epsilon_i a_i^\dagger a_i - \frac{g}{4} \sum_{ij} a_i^\dagger a^\dagger_{\bar{i}} a_{\bar{j}} a_j\]

One can use the following methods to solve the pairing Hamiltonian with the package:

"Full-CI": Full-CI/Exact diagonalization

"Full-CI(2-fold)": Full-CI/Exact diagonalization assuming that particles are paired

"HF": Hartree-Fock (nothing but naive filling due to the pairing Hamiltonian) Under this method, MBPT calculation with the HF reference state are carried out.

"BCS": Bardeen-Cooper-Schrieffer

"CCD": Coupled Cluster, especially CCD

"IMSRG(2)": In-Medium Similarity Renormalization Group, especially IM-SRG(2)

PairingHamiltonians.PairingHamiltonian — Method

PairingHamiltonian(;Norb_in::Int=8, Nocc_in::Int=4, gval::Float64=0.33, delta_eps::Float64=1.0,
                    debug_mode::Int=0, solver::String="FCI(2-fold)", save_Exact_wf::Bool=false, to_in=nothing)

Main function to evaluate the ground state energy of the pairing Hamiltonian.

Optional arguments with default values

Norb_in::Int=8: number of orbitals

Nocc_in::Int=4: number of occupied states

gval::Float64=0.33: pairing strength

delta_eps::Float64=1.0: energy difference between orbitals

debug_mode::Int=0: specify the debug mode

solver::String("FCI(2-fold)"): method to solve the Hamiltonian. This can be one of "Full-CI", "Full-CI(2-fold)", "HF", "BCS", "CCD", "IMSRG(2)". If "HF" is chosen, PT2/PT3 energies are also calculated.

save_Exact_wf::Bool(false): save the full-CI wave functions as HDF5 files, which can be used for e.g. analysis of the wave functions and constructing surrogate models like eigenvector continuation.

to: either nothing or TimerOutput object defined in a user script.

PairingHamiltonians.bitstr2int — Method

bitstr2int(bitstr::String)::Int

Convert a bitstring representation (e.g., "00110") to an integer (e.g., 6).

PairingHamiltonians.check_connectivity — Method

check_connectivity(psi_l::Int, psi_r::Int, p::Int, q::Int, Nq::Int)::Bool

Check if the configurations (in bit representation) are connected by the Hamiltonian.

PairingHamiltonians.eval_Hamil — Method

eval_Hamil(method::String, basis::Vector{Int}, eps::Vector{Float64}, g::Float64, Nq::Int, degenerate::Bool; debug_mode::Int=0, sparce_rep::Bool=false)::Tuple{Matrix{Float64}, Matrix{Float64}, Union{Matrix{Float64}, Dict{UInt64, Float64}}}

Function to evaluate the Hamiltonian matrix elements. Once the number of orbitals and the number of occupied states are given, one can prepare the Hamiltonian matrix elements in the full matrix representation or sparse representation. For larger systems, sparse representation using Dict{UInt64, Float64} is used instead of storing the full matrix. The keys of the dictionary correspond to the integer representation of many-body configurations.

Note that "1" and '1' are different in Julia. The former is a string, while the latter is a character.

PairingHamiltonians.hash_key2 — Method

hash_key2(i::Int, j::Int)::UInt64

Hash the pair of indices to a key.

PairingHamiltonians.int2bitstr — Method

int2bitstr(x::Int, N::Int)::String

Convert an integer to a bitstring of length N. Similar thing can be done with string function in Julia, but this is more proper way for "0" padding.

PairingHamiltonians.prepare_basis — Method

prepare_basis(N::Int, n::Int)::Vector{Int}

Create a list of configurations where n out of N are occupied.

PairingHamiltonians.unhash_key2 — Method

unhash_key2(key::UInt64)::Tuple{Int, Int}

Unhash the key to the pair of indices.