Hartree-Fock

The Hartree-Fock (HF) method is based on the assumption that the many-body wave function can be approximated by a single Slater determinant. The HF is nothing but a naive filling for the global pairing Hamiltonian, but it is a good starting point for more advanced methods such as MBPT, post-HF methods (CC, IMSRG, etc.). For this reason, some functions are not necessary for the HF method, but they are implemented for future use.

PairingHamiltonian.HF — Function

HF(Nocc, h1b, h2b, gval, to, debug_mode=0; itnum_max=100, tol=1e-9)

Main function to compute the energies of the Hamiltonian by the Hartree-Fock method and perturbation theory.

Arguments

Nocc::Int64: Number of occupied orbitals.

h1b::Matrix{Float64,2}: One-body Hamiltonian.

h2b::Vector{Float64,1}: Two-body Hamiltonian flattened. This will be used only for the normal ordering.

gval::Float64: Interaction strength.

to::TimeOutput: TimeOutput object to measure the time.

Optional arguments

debug_mode::Int64(0): If 1, print the energies at each iteration.

itnum_max::Int64(100): Maximum number of iterations.

tol::Float64(1e-9): Tolerance to stop the iteration.

PairingHamiltonian.define_holes_particles — Method

define_holes_particles(F, Nocc)

Function to return the indices of the holes, which are below the Fermi level, and the particles, which are above the Fermi level.

PairingHamiltonian.eval_EHF — Method

eval_EHF(h1b, gval, rho, Nocc)

Function to evaluate the energy of the Hamiltonian in the Hartree-Fock method.

\[\begin{align} E &= E_{1b} + E_{2b} \nonumber \\ E_{1b} & = \sum_{i \leq F} \sum_{\alpha\beta} C^T_{i \alpha} C_{i \beta} \langle \alpha | h | \beta \rangle = \sum_{\alpha \beta} \rho_{\alpha \beta} \langle \alpha | h | \beta \rangle\nonumber \\ E_{2b} &= \frac{1}{2}\sum_{\alpha\beta\gamma\delta} \rho_{\alpha \beta} \rho_{\gamma \delta} \langle \alpha \gamma | v | \beta \delta \rangle \nonumber \end{align}\]

PairingHamiltonian.eval_Fock! — Method

eval_Fock!(F, rho, h1b, Nocc, gval)

Function to evaluate the Fock matrix destructively.

\[F_{\alpha\beta} = \langle \alpha | h | \beta \rangle + \sum_{\gamma \delta} \rho_{\gamma \delta} \langle \alpha \gamma | v | \beta \delta \rangle\]

PairingHamiltonian.eval_rho! — Method

eval_rho!(rho, U, Nocc)

Update the density matrix with the coefficient matrix.

\[\rho_{\alpha\beta} = \sum_{i \leq F} C^T_{i \alpha} C_{i \beta}\]