In-Medium Similarity Renormalization Group (IMSRG)

Functions for the In-Medium Similarity Renormalization Group (IMSRG) method.

Under the current implementation, the IMSRG is the most time-consuming method among the ones available in the package. One should consider to run a script using multiple threads to speed up the calculations like the following example:

julia -t 12 path_to_script.jl

Some functions in the package are already parallelized, but it may be still insufficient for large systems.

PairingHamiltonians.IMSRG — Method

IMSRG(HNO, holes, particles, gval, Nocc, to, debug_mode; smax = 15.0, tol_eta=1.e-4, ds=1.e-2)

Main function to perform the IMSRG calculation.

Arguments

HNO::Hamiltonian_NormalOrdered: Hamiltonian in normal ordered form

holes::Vector{Int64}: indices of the holes

particles::Vector{Int64}: indices of the particles

gval::Float64: strength of the interaction

Nocc::Int64: number of occupied orbitals

to::TimeOutput: output object to measure the time

debug_mode::Int64: level of debug information

Optional arguments

smax::Float64: maximum value of the flow parameter

tol_eta::Float64: tolerance of the generator

ds::Float64: step size of the flow parameter

References

Heiko Hergert, Scott K. Bogner, Justin G. Lietz, Titus D. Morris, Samuel J. Novario, Nathan M. Parzuchowski & Fei Yuan: In-Medium Similarity Renormalization Group Approach to the Nuclear Many-Body Problem, part of the book series: Lecture Notes in Physics ((LNP,volume 936))

PairingHamiltonians.eta_white_atan! — Method

eta_white_atan!(eta_1b, eta_2b, f, Γ, idx2b, kets, holes, particles)

Destructive function to evaluate the generator of IMSRG flow equation, η.

\[\begin{align} \eta & = \eta^{(1)} + \eta^{(2)} \nonumber \\ & = \frac{1}{2} \sum_{ph} \arctan\left(\frac{f_{ph}(s)}{\Delta_{ph}(s)}\right) \{ a^\dagger_p a_h \} + \frac{1}{8} \sum_{pp'hh'} \arctan\left(\frac{2 \Gamma_{pp'h'h}(s)}{\Delta_{pp'hh'}(s)}\right) \{ a^\dagger_p a^\dagger_{p'} a_h a_{h'} \} - H.c.. \nonumber \end{align}\]

PairingHamiltonians.eval_2b_ph_trans! — Method

eval_2b_ph_trans!(eta, eta_ph, Γ, Γ_ph, idx2b, kets)

Evaluating $A_{i\bar{j}k\bar{l}} = - A_{ijkl}$ destructively. Those will be used for making the 2-body part of the flow equation faster.

PairingHamiltonians.eval_derivatives! — Method

eval_derivatives!(dH, eta_1b, eta_2b, f, Γ, idx2b, coppy_matrices, kets, kets_ph,                     
                       holes, particles, holes_2b, particles_2b, to, debug_mode)

To evaluate the derivatives of the Hamiltonian operator, we need to evaluate the following terms:

\[\begin{align} \frac{dE}{ds} &= \sum_{ab} (n_a-n_b) \eta^{(1)}_{ab} f_{ba} + \frac{1}{2} \sum_{abcd} \eta^{(2)}_{abcd} \Gamma_{cdab} n_an_b \bar{n}_c \bar{n}_d \nonumber \\ \frac{df_{pq}}{ds} &= \sum_{a} (1+P_{pq}) \eta^{(1)}_{pa} f_{aq} + \sum_{ab} (n_a-n_b) (\eta^{(1)}_{ab} \Gamma_{bpaq} - f_{ab} \eta^{(2)}_{bpaq}) \nonumber \\ & + \frac{1}{2} \sum_{abc} (n_an_b\bar{n}_c + \bar{n}_a \bar{n}_b n_c) (1+P_{pq}) \eta^{(2)}_{cpab} \Gamma_{abcq} \nonumber \\ \frac{d\Gamma_{pqrs}}{ds} &= \sum_{a} \left\{ (1-P_{pq}) (\eta^{(1)}_{pa} \Gamma_{aqrs} - f_{pa} \eta^{(2)}_{aqrs}) -(1-P_{rs}) (\eta^{(1)}_{ar} \Gamma_{pqas} - f_{ar} \eta^{(2)}_{pqas}) \right\} \nonumber \\ & + \frac{1}{2} \sum_{ab} (1-n_a-n_b) (\eta^{(2)}_{pqab} \Gamma_{abrs} - \Gamma_{pqab} \eta^{(2)}_{abrs})\nonumber \\ & - \sum_{ab} (n_a - n_b) (1-P_{pq})(1-P_{rs}) \eta^{(2)}_{bqas} \Gamma_{apbr} \nonumber \end{align}\]

where $P_{pq}$ is the permutation operator, and $n_a$ is the occupation number of the orbital $a$.