Many-Body Perturbation Theory

One can use the Many-Body Perturbation Theory (MBPT) to calculate the ground state energy of a system. In the package, PT2 and PT3 are implemented, and can be used to calculate the ground state energy of a system. Users do not have to specify "MBPT2" or "MBPT3" in the input file, as the package will automatically use the methods when the Hartree-Fock calculation is called.

PairingHamiltonian.PT2 — Method

PT2(F, Nocc, gval)

Function to calculate the second order perturbation theory contribution to the energy.

\[\frac{1}{4} \sum^{\leq \epsilon_F}_{ij}\sum^{\geq \epsilon_F}_{pq} \frac{ \langle ij | V | pq \rangle \langle pq | V | ij \rangle }{\epsilon_i + \epsilon_j - \epsilon_p - \epsilon_q} \]

For the pairing Hamiltonian, j should be the pair of i and p should be the pair of q, and the numerator will be g^2.

PairingHamiltonian.PT3 — Method

PT3(F, Nocc, gval)

Function to calculate the third order perturbation theory contribution to the energy.

\[\begin{align} \Delta E^{(3)}_{pp} & = \frac{1}{8} \sum_{ij \leq F} \sum_{abcd > F} \frac{\bra{ij} \hat{V} \ket{ab} \bra{ab} \hat{V} \ket{cd} \bra{cd} \hat{V} \ket{ij}}{ \epsilon^{ab}_{ij} \epsilon^{cd}_{ij}} \\ \Delta E^{(3)}_{hh} & = \frac{1}{8} \sum_{ijkl \leq F} \sum_{ab > F} \frac{\bra{ij} \hat{V} \ket{ab} \bra{kl} \hat{V} \ket{kl} \bra{kl} \hat{V} \ket{ij}}{ \epsilon^{ab}_{ij} \epsilon^{ab}_{kl}} \\ \Delta E^{(3)}_{ph} & = - \sum_{ijk \leq F} \sum_{abc > F} \frac{\bra{ij} \hat{V} \ket{ab} \bra{kb} \hat{V} \ket{ic} \bra{ac} \hat{V} \ket{kj}}{ \epsilon^{ab}_{ij} \epsilon^{ac}_{kj}} \end{align}\]

For the global pairing Hamiltonian, the PT3 is much simpler than the general case as in the PT2, since the numerator will be proportional to g^3 and the ph term is zero.