Coupled Cluster

Functions for performing coupled cluster calculations:

\[\begin{align*} \ket{\Psi_\mathrm{CC}} & = e^{\hat{T}} \ket{\Phi_0}\\ \hat{T} & = \hat{T}_1 + \hat{T}_2 + \hat{T}_3 + \cdots + \hat{T}_A \\ \hat{T}_n & = \left( \frac{1}{n!} \right)^2 \sum^{a_1 a_2 \ldots a_n}_{i_1 i_2 \ldots i_n} t^{a_1 a_2 \ldots a_n}_{i_1 i_2 \ldots i_n} a^\dagger_{a_1} a^\dagger_{a_2} \ldots a^\dagger_{a_n} a_{i_n} a_{i_{n-1}} \ldots a_{i_1} \end{align*}\]

Since the current Hamiltonian is a pairing Hamiltonian, we will only consider the CCD method, $\hat{T} = \hat{T}_2$.

PairingHamiltonians.CCD — Method

CCD(F, gval, Nocc, to, debug_mode; itnum=100, tol=1e-9, mix_ratio=0.25)

Main function to calculate the ground state energy of the pairing Hamiltonian using the coupled cluster method. Since the Hamiltonian is the pairing model, we don't need to consider 1p1h excitations, CCD instead of CCSD, etc.

Arguments

F::Array{Float64, 2}: Fock matrix from the Hartree-Fock calculation

gval::Float64: strength of the pairing interaction

Nocc::Int: number of occupied states

to::TimerOutput: timer object to measure the elapsed time

debug_mode::Int: specifying the debug mode

Optional arguments

itnum::Int(100): maximum number of iterations

tol::Float64(1e-9): convergence criterion

mix_ratio::Float64(0.25): mixing ratio for the update of the t2 amplitudes

References

Shavitt, I., & Bartlett, R. J. (2009). Many-body methods in chemistry and physics: MBPT and coupled-cluster theory. Cambridge University Press.

Justin G. Lietz, Samuel Novario, Gustav R. Jansen, Gaute Hagen & Morten Hjorth-Jensen, Computational Nuclear Physics and Post Hartree-Fock Methods, part of the book series: Lecture Notes in Physics ((LNP,volume 936))

PairingHamiltonians.construct_chi — Function

construct_chi(F, t2_pphh, Nocc, gval, debug_mode=0)

Function to construct the intermediate matrix $\chi$ for the CCD calculation.

\[\begin{align*} \langle b | \chi | c \rangle & = \langle b | f | c \rangle - \frac{1}{2} \sum_{kld} \langle bd | t | kl \rangle \langle kl | v | cd \rangle, \\ \langle k | \chi | j \rangle & = \langle k | f | j \rangle + \frac{1}{2} \sum_{cdl} \langle kl | v | cd \rangle \langle cd | t | jl \rangle, \\ \langle kl | \chi | ij \rangle & = \langle kl | v | ij \rangle + \frac{1}{2} \sum_{cd} \langle kl | v | cd \rangle \langle cd | t | ij \rangle, \\ \langle kb | \chi | cj \rangle & = \langle kb | v | cj \rangle + \frac{1}{2} \sum_{ld} \langle kl | v | cd \rangle \langle db | t | lj \rangle, \\ \langle ab | \chi | cd \rangle &= \langle ab | v | cd \rangle. \end{align*}\]

where i,j,k are hole indices, a,b,c are particle indices.

PairingHamiltonians.eval_ECC — Method

eval_ECC(F, t2_pphh, Nocc, gval)

Function to evaluate the CCD energy.

\[E_{\mathrm{CC}} = \frac{1}{4}\sum_{ijab} \langle ab | \hat{v} | ij \rangle t^{ab}_{ij}\]

PairingHamiltonians.eval_Hbar_pphh! — Method

eval_Hbar_pphh!(Hbar_pphh, t2_pphh, t2_pphh_new, F, χ::chi_intermediate, Nocc, gval)

Function to evaluate the Hbar operator for the CCD calculation and update the t2 amplitudes.

\[\begin{align*} t^{ab \mathrm{(new)}}_{ij} = \frac{1}{\epsilon^{ab}_{ij}} & \left\{ \langle ab | \hat{v} | ij \rangle +\hat{P}(ab) \sum_c \langle b | \chi | c \rangle t^{ac}_{ij} -\hat{P}(ij) \sum_k \langle k | \chi | j \rangle t^{ab}_{ik} \right. \\ & \left. + \frac{1}{2} \sum_{cd} \langle ab | \chi | cd \rangle t^{cd}_{ij} + \frac{1}{2} \sum_{kl} \langle kl | \chi | ij \rangle t^{ab}_{kl} + \hat{P}(ij) \hat{P}(ab) \sum_{kc} \langle kb | \chi | cj \rangle t^{ac}_{ik} \right\} \end{align*}\]

To avoid being stuck, t2 values will be updated by the following mixing rule:

\[t^{ab \mathrm{(new)}}_{ij} = \alpha t^{ab \mathrm{(new)}}_{ij} + (1-\alpha) t^{ab}_{ij}\]

where $\alpha$ is the mix_ratio specified in CCD function.

PairingHamiltonians.eval_χ_hhhh! — Method

Corresponding to the Eq.(8.48) of LNP volume 936.

\[\langle kl | \chi | ij \rangle = \langle kl | v | ij \rangle + \frac{1}{2} \sum_{cd} \langle kl | v | cd \rangle \langle cd | t | ij \rangle\]

PairingHamiltonians.eval_χ_hphp! — Method

Corresponding to the Eq.(8.49) of LNP volume 936.

\[\langle kb | \chi | cj \rangle = \langle kb | v | cj \rangle + \frac{1}{2} \sum_{ld} \langle kl | v | cd \rangle \langle db | t | lj \rangle,\]

PairingHamiltonians.eval_χ_pppp! — Method

Corresponding to the Eq.(8.50) of LNP volume 936.

\[\langle ab | \chi | cd \rangle = \langle ab | v | cd \rangle\]

PairingHamiltonians.hash_key4 — Method

hash_key4(i, j, a, b)::UInt64

Function to encode i,j,a,b into a single UInt64. From the lowest 8 bits, b, a, j, i are stored. If you want to work with larger model space, requiring e.g. single particle states > 255, you need to modify this function.

PairingHamiltonians.unhash_key4 — Method

unhash_key4(key::UInt64)::Tuple{Int, Int, Int, Int}

Function to decode i,j,a,b from a single UInt64.