Full-CI

PairingHamiltonian.FCI — Function

FCI(Hamil_mat, dim_basis, Norb, Nocc, gval, save_Exact_wf, to, debug_mode=0; return_only_E0=true)

Main function to compute the ground state energy with the full CI method. If the system size is small, the Hamiltonian matrix is explicitly constructed and diagonalized using the eigen function in LinearAlgebra.jl. If the system size is large, the Hamiltonian matrix is sparsely represented and the Lanczos method is used to compute the smallest eigenvalue.

Arguments

Hamil_mat: Hamiltonian matrix, either a Matrix{Float64} or a Dict{UInt64, Float64}.

dim_basis: Dimension of the basis, i.e. number of configurations

Norb: Number of orbitals

Nocc: Number of occupied orbitals

gval: Interaction strength

save_Exact_wf: If true, save the exact wave function to a HDF5 file

to: TimerOutput object

debug_mode: the level of debug information

return_only_E0: If true, return only the ground state energy, otherwise return the eigenvalues and eigenvectors

PairingHamiltonian.lanczos — Method

lanczos(Hamil_mat, dim, save_Exact_wf, to; itnum=300, tol=1e-9, debug_mode=0)

Function to compute the lowest eigenvalue of the Hamiltonian using the Lanczos method.

Constructing a Krylov subspace $\mathcal{K}_m(H,v) = \mathrm{span}\{v, Hv, H^2v, \cdots, H^{m-1}v\}$, and the tridiagonal matrix $T_m = V_m^T H V_m$ where $V_m = [v_1, v_2, \cdots, v_m]$ and $v_{m+1} = H v_m - \alpha_m v_m - \beta_m v_{m-1}$, the Lanczos method iteratively constructs the matrix $T_m$ and diagonalizes it to obtain the smallest eigenvalue of $H$.

Arguments

Hamil_mat: Hamiltonian matrix, either a Matrix{Float64} or a Dict{UInt64, Float64}.

dim: Dimension of the basis, i.e. number of configurations

save_Exact_wf: If true, save the exact wave function to a HDF5 file

to: TimerOutput object

Optional arguments

itnum: Number of Lanczos iterations

tol: Tolerance for convergence

debug_mode: the level of debug information

PairingHamiltonian.make_gs_wf — Method

make_gs_wf(vks, Tmat, Lan_itnum, num_ev)

Function to construct the ground state wave function from the Lanczos vectors.

\[|\psi_{\rm GS}\rangle = \sum_{k=1}^{D} c_k |v_k\rangle\]

where D is the number of Lanczos iterations, and $c_k$ is the eigenvector of the tridiagonal matrix $T$ corresponding to the smallest eigenvalue.

Arguments

vks: Lanczos vectors

Tmat: Tridiagonal matrix

Lan_itnum: Number of Lanczos iterations

num_ev: Number of the eigenvector to be used

PairingHamiltonian.operate_H_on_vec! — Method

operate_H_on_vec!(w, Hamil_mat::Dict{UInt64, Float64}, v)

Sparse version of the function to compute the matrix-vector product of the Hamiltonian matrix in the form of Dict{UInt64, Float64} and a vector.

PairingHamiltonian.operate_H_on_vec! — Method

operate_H_on_vec!(w, Hamil_mat::Array{Float64, 2}, v)

Function to compute the matrix-vector product of the Hamiltonian matrix and a vector.

PairingHamiltonian.reOrthogonalize! — Method

reOrthogonalize!(w, vks, i)

Re-orthogonalize the vector w with respect to the previous vectors:

\[w := w - \sum_{j=1}^{i} \langle w, v_j \rangle v_j\]

PairingHamiltonian.write_Exact_wf_hdf5 — Method

write_Exact_wf_hdf5(evec, Norb, Nocc, gval)

Function to save the exact wave function to a HDF5 file.

Arguments

evec: Exact wave function

Norb: Number of orbitals

Nocc: Number of occupied orbitals

gval: Interaction strength