In-Medium Similarity Renormalization Group (IMSRG)

Functions for the In-Medium Similarity Renormalization Group (IMSRG) method.

Under the current implementation, the IMSRG is the most time-consuming method among the ones available in the package. One should consider to run a script using multiple threads to speed up the calculations like the following example:

julia -t 12 path_to_script.jl

Some functions in the package are already parallelized, but it may be still insufficient for large systems.

PairingHamiltonian.IMSRG — Method

IMSRG(HNO, holes, particles, gval, Nocc, to, debug_mode;
            smax = 15.0, tol_eta=1.e-4, ds=1.e-2)

Main function to perform the IMSRG calculation.

Arguments

HNO::Hamiltonian_NormalOrdered: Hamiltonian in normal ordered form

holes::Vector{Int64}: indices of the holes

particles::Vector{Int64}: indices of the particles

gval::Float64: strength of the interaction

Nocc::Int64: number of occupied orbitals

to::TimeOutput: output object to measure the time

debug_mode::Int64: level of debug information

Optional arguments

smax::Float64: maximum value of the flow parameter

tol_eta::Float64: tolerance of the generator

ds::Float64: step size of the flow parameter

PairingHamiltonian.eta_white_atan! — Method

eta_white_atan!(eta_1b, eta_2b, f, Γ, idx2b, kets, holes, particles)

Destructive function to evaluate the generator of IMSRG flow equation, η.

\[\begin{align} \eta & = \eta^{(1)} + \eta^{(2)} \nonumber \\ & = \frac{1}{2} \sum_{ph} \arctan\left(\frac{f_{ph}(s)}{\Delta_{ph}(s)}\right) \{ a^\dagger_p a_h \} + \frac{1}{8} \sum_{pp'hh'} \arctan\left(\frac{2 \Gamma_{pp'h'h}(s)}{\Delta_{pp'hh'}(s)}\right) \{ a^\dagger_p a^\dagger_{p'} a_h a_{h'} \} - H.c.. \nonumber \end{align}\]

PairingHamiltonian.eval_2b_ph_trans! — Method

Evaluating $A_{i\bar{j}k\bar{l}} = - A_{ijkl}$